Refactor controllers
Separate chat and chatThread.
Attempt to "extract method" for getting context and basic validation.
